refactor: comment mutation 파일 분리
